wip5 - trainer code structure prettifying
