+++ app/trainer/__init__.py
@@ -1,0 +1,7 @@
+from .pipelines import LRPipeline, NNPipeline, TreePipeline
+
+__all__ = (
+    'LRPipeline',
+    'NNPipeline',
+    'TreePipeline'
+)
