+++ app/trainer/pipelines/__init__.py
@@ -1,0 +1,9 @@
+from .lr_pipeline import LRPipeline
+from .nn_pipeline import NNPipeline
+from .tree_pipeline import TreePipeline
+
+__all__ = (
+    'LRPipeline',
+    'NNPipeline',
+    'TreePipeline'
+)
