+++ app/trainer/transformer/__init__.py
@@ -1,0 +1,5 @@
+from .transformer import Transformer
+
+__all__ = (
+    'Transformer'
+)
